Initial work to parse profile into JSON via a feature
The idea behind this is that I can commit the JSON for
my build script instead of the binary excel file. When we
need to process the profile excel file then we can compile with
the feature and parse the excel file.

This will not work if the JSON generation and parsing isn't done
in a repeatable fashion since then my module code will have random
changes.