Remove separate profile parser code
I am merging this back into the build script and
will simply commit the profile to the repo since
that is the easiest option and most sensical. The
file is small and should have a minimal impact on
repo size. My JSON dump was one giant mini-fied line
and changes to it would also likely create very large
plain text diffs all the same.